embedding-pipeline/requirements.txt
anyio==3.7.0
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dill==0.3.6
docutils==0.20.1
exceptiongroup==1.1.1
fastapi==0.96.1
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
keyring==23.13.1
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
minio==7.1.15
mmh3==3.0.0
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
onnx==1.14.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkginfo==1.9.6
pluggy==1.0.0
protobuf==4.23.2
psutil==5.9.5
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pymilvus==2.1.1
PyMySQL==1.0.3
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
readme-renderer==37.3
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
towhee==1.0.0
towhee.models==1.0.0
tqdm==4.65.0
transformers==4.30.0
twine==4.0.2
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.4.0
ultralytics==8.0.114
urllib3==2.0.3
uvicorn==0.22.0
webencodings==0.5.1
zipp==3.15.0
